requirements.txt
Flask==2.3.3
pyodbc==4.0.39
python-dotenv==1.0.0